lib/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
  packages = ["."]

[project.scripts]
olivier = "olivier.olivier:app"

[project]
name = "olivier"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
  "cachetools==5.3.1",
  "dateparser==1.2.2",
  "grist-api==0.1.1",
  "python-dateutil==2.8.2",
  "requests==2.31.0",
  "typer==0.24.1",
  "Unidecode==1.3.6",
]
